move emails strategy to public API. move emails strategy from provisional to strategies.py.
add entry to CONTRIBUTING.rst

fix lint and check format issue on travis.
